Addressed code review feedback.
