+++ src/Umbraco.Web.UI.Client/src/packages/dynamic-root/index.ts
@@ -1,0 +1,3 @@
+export * from './dynamic-root-resolver.controller.js';
+export * from './repository/index.js';
+export type * from './types.js';
