+++ src/Umbraco.Web.UI.Client/src/packages/media/media/property-editors/media-picker/Umbraco.SingleMediaPicker.ts
@@ -40,7 +40,7 @@ export const manifest: ManifestPropertyEditorSchema = {
 				{
 					alias: 'ignoreUserStartNodes',
 					label: 'Ignore User Start Nodes',
-					description: 'Selecting this option allows a user to choose nodes that they normally dont have access to.',
+					description: "Selecting this option allows a user to choose nodes that they normally don't have access to.",
 					propertyEditorUiAlias: 'Umb.PropertyEditorUi.Toggle',
 				},
 			],
