+++ src/Umbraco.Web.UI.Client/src/packages/multi-url-picker/property-editor/Umbraco.SingleUrlPicker.ts
@@ -11,7 +11,7 @@ export const manifest: ManifestPropertyEditorSchema = {
 				{
 					alias: 'ignoreUserStartNodes',
 					label: 'Ignore user start nodes',
-					description: 'Selecting this option allows a user to choose nodes that they normally dont have access to.',
+					description: "Selecting this option allows a user to choose nodes that they normally don't have access to.",
 					propertyEditorUiAlias: 'Umb.PropertyEditorUi.Toggle',
 				},
 			],
