+++ src/Umbraco.Web.UI.Client/src/packages/property-editors/content-picker/dynamic-root/index.ts
@@ -1,5 +1,3 @@
 export * from './modals/index.js';
-export * from './repository/index.js';
 export * from './components/index.js';
-export * from './dynamic-root-resolver.controller.js';
 export type * from './dynamic-root.extension.js';
