+++ src/Umbraco.Web.UI.Client/src/packages/property-editors/content-picker/index.ts
@@ -2,4 +2,7 @@ export * from './components/index.js';
 export * from './config/source-content/index.js';
 export * from './constants.js';
 export * from './dynamic-root/index.js';
+// Kept exported from here as well: the repository and resolver were public API of this package before they
+// moved into their own module.
+export { UmbContentPickerDynamicRootRepository, UmbDynamicRootResolver } from '@umbraco-cms/backoffice/dynamic-root';
 export type * from './types.js';
