+++ src/Umbraco.Web.UI.Client/src/packages/property-editors/content-picker/types.ts
@@ -1,4 +1,9 @@
+import type { UmbContentPickerDynamicRoot } from '@umbraco-cms/backoffice/dynamic-root';
+
 export type * from './dynamic-root/types.js';
+// The dynamic root types now live in their own module, so the pickers that offer one need not depend on the
+// content picker. Re-exported here because the names are public API.
+export type * from '@umbraco-cms/backoffice/dynamic-root';
 
 export type UmbContentPickerSourceType = 'content' | 'member' | 'media';
 
@@ -7,15 +12,3 @@ export type UmbContentPickerSource = {
 	id?: string;
 	dynamicRoot?: UmbContentPickerDynamicRoot;
 };
-
-export interface UmbContentPickerDynamicRoot {
-	originAlias: string;
-	originKey?: string;
-	querySteps?: Array<UmbContentPickerDynamicRootQueryStep>;
-}
-
-export interface UmbContentPickerDynamicRootQueryStep {
-	unique: string;
-	alias: string;
-	anyOfDocTypeKeys?: Array<string>;
-}
