+++ src/Umbraco.Web.UI.Client/tsconfig.json
@@ -72,6 +72,7 @@ DON'T EDIT THIS FILE DIRECTLY. It is generated by /devops/tsconfig/index.js
 			"@umbraco-cms/backoffice/dictionary": ["./src/packages/dictionary/index.ts"],
 			"@umbraco-cms/backoffice/document-blueprint": ["./src/packages/documents/document-blueprints/index.ts"],
 			"@umbraco-cms/backoffice/document-type": ["./src/packages/documents/document-types/index.ts"],
+			"@umbraco-cms/backoffice/dynamic-root": ["./src/packages/dynamic-root/index.ts"],
 			"@umbraco-cms/backoffice/document": ["./src/packages/documents/documents/index.ts"],
 			"@umbraco-cms/backoffice/dropzone": ["./src/packages/media/dropzone/index.ts"],
 			"@umbraco-cms/backoffice/entity-action": ["./src/packages/core/entity-action/index.ts"],
